Moving interfaces to same package
