+++ onecalendar/src/main/java/com/darwindeveloper/onecalendar/domain/OnClickDayListener.java
@@ -1,4 +1,4 @@
-package com.darwindeveloper.onecalendar.list;
+package com.darwindeveloper.onecalendar.domain;
 
 import com.darwindeveloper.onecalendar.model.Day;
